REFACT/MAJOR: Separated logic between SWAN access point, CMP, Publisher, Marketer and OpenRTB processors in the demo.
Major refactoring of the code to improve readability, reuse and clear logic ready for improved documentation.

Shifted the stop, preference and info user interfaces into the CMP part of the demo. SWAN is now merely a data access layer. SWIFT does have some user interface components for the progress indicator and messages concerning the configuration of the web browser.

The decoding of storage operation via SWAN now have two endpoints.

/swan/api/v1/values-as-json – checks the expiry time of the data and returns an error if the data has expired. This end point returns the hashed email address and is the one that should be used for read operations for the purposes of advertising where the original email address is not needed.

/swan/api/v1/operation-as-json – does not check the expiry time and is used by the CMP to obtain the storage operation parameters from the URL path when commencing the second stage of the storage operation after the user has edited the data. This does return the email address in it’s original form for display in the user interface. It should only be used by CMPs for editing data and never to return the email address for use in an advertising transaction. This would constitute a violation of the SWAN model contract clauses.

Modified the SWIFT layer to enable SWAN Operator domains that are also SWIFT access nodes to avoid making HTTP calls. Improves performance.

Modified the SWAN access control to check for headers from web browsers that would not typically be present in server to server requests. If these headers are present then access control will fail. This is a safety check to reduce the risk of a SWAN access key becoming  known to an unauthorized party.

Modified the SWAN project to reduce the number of browser redirects required.

Added gzip compression to the HTTP responses.