FEAT/CLEANUP: Removed the dialog handler. This is now accomplished via Fetch using the state array in the SWIFT storage operation.
Access Key is now removed from the list of parameters after it has been used. This will help prevent it in advertently being passed to other methods. Fixed a bug in Stop where the access key was being added from the old configuration.

Tidy up of some comments and code.