FEAT/BUG: Callers to SWAN can now optionally pass existing SWID and preference parameters for fetch actions. Minor bugs.
If the preference value is a valid OWID and has not expired, then it will be returned from the fetch operation if no other value exists.

If the SWID value is also a valid OWID AND it was created by a SWAN Operator know to the processing SWAN Operator by virtue of access node knowledge the SWID value will be used if no other value exists.

This enhancement enables participating publishers to use their domains to store copies of SWAN data, not only for their own purposes, but also for maintain the SWAN network when tracking prevention techniques are removing cookies for the SWAN Operator domains. Publishers who do not wish to participate in this was do not have to. The option now exists.

An additional parameter is returned from fetch requests to the SWAN network to indicate when the values returned should be revalidated with the SWAN network. Previously the expiry date was being used in the SWAN key value pair to indicate when the cookie should expire. This was too short a period of time and removed meaning from the data. The expiry time is used elsewhere to indicate when the value should be removed from the SWAN network. This is now set consistently throughout the SWAN demo. This change enables callers to persist the SWAN data over a longer period of time and to then use the revalidate parameter to determine when to revalidate with SWAN in case the data has changed by virtue of a request to another domain.

The swan.Pair now has a value that is a string and not a byte array. As the structure is more commonly used to record cookies and pass values in POST requests this reduces the need to turn data from byte arrays to strings.

The parameters needed to identify the home node were not being set for all operations. This has now been resolved.

The storage operation timeout was sometimes too short for the increased number of nodes used in the demo for the stop ads feature. This would result in the UIP dialog appearing when it shouldn’t.

The stopped domains are now space separated to ensure the returned string complies with cookie value policies.

SWAN cookies are now prefixed “swan-“. This should avoid naming conflict with other cookies used by the caller. This is not a SWAN rule but one that will follow from the demo.

Added a new method to the swan.Pair to return the key as a cookie name with the prefix.